% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxknapsack.R
\docType{package}
\name{boxknapsack}
\alias{boxknapsack}
\alias{boxknapsack-package}
\title{boxknapsack: A package use three different to solve knapsack problem.}
\description{
A package use three different to solve knapsack problem.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
\author{
Maintainer: Boxi Zhang bossyzhang@gmail.com
}
