% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_para.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{is a matrix containing the weights and values}

\item{W}{a numeric string.}

\item{parallel}{is the logical}
}
\value{
a list with the _value_ of the optimally packed knapsack and the _elements_ that gives this value.
}
\description{
tests all combinations and finds the one with max value under the restriction total weight<W.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
