% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a matrix containing the weights and values}

\item{W}{a numeric string.}
}
\value{
A list cointaining a value and a element
}
\description{
The 'greedy_knapsack' function is a approach that use the a heuristic or approximation for knapsack package problem.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
 )
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
