% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{is a matrix containing the weights and values}

\item{W}{a numeric string.}
}
\value{
A list cointaining a value and a element
}
\description{
The 'knapsack_dynamic' function is a approach that use the a heuristic or approximation for knapsack package problem#' @param x is a 2dim matrix containing the weights and values.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
 )
knapsack_dynamic(x = knapsack_objects[1:800,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
