% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is a matrix containing the weights and values}

\item{W}{a numeric string.}
}
\value{
a list with the _value_ of the optimally packed knapsack and the _elements_ that gives this value.
}
\description{
tests all combinations and finds the one with max value under the restriction total weight<W.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <-
 data.frame(
   w=sample(1:4000, size = n, replace = TRUE),
  v=runif(n = n, 0, 10000)
 )
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
